+++ apps/openmw/mwbase/journal.hpp
@@ -74,7 +74,7 @@ namespace MWBase
 
         virtual const TQuestContainer& getQuests() const = 0;
 
-        virtual int countSavedGameRecords() const = 0;
+        virtual size_t countSavedGameRecords() const = 0;
 
         virtual void write(ESM::ESMWriter& writer, Loading::Listener& progress) const = 0;
